rename java-packet (hw.zako.multichat.packet -> hw.zako.multichat.events)
